refactor: extract a run hook, step row, and types from the apply drawer
Move the auto-apply job lifecycle (start/cancel + job-event step/outcome tracking)
into useApplyRun, the step renderer into StepRow, and the Posting/ApplyStep
interfaces into types.ts. index.tsx keeps the form inputs and layout.
